pyproject.toml
[project]
name = "mempool-monitor"
version = "0.1.0"
description = "Mempool scanner for Uniswap V2/V3 swaps on Ethereum"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "web3>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
